+++ R/trace.R
@@ -371,10 +371,10 @@ trace_length <- function(x) {
 }
 
 trace_subset <- function(x, i) {
-  stopifnot(is_integerish(i))
   if (!length(i)) {
     return(new_trace(list(), int(), list()))
   }
+  stopifnot(is_integerish(i))
 
   n <- trace_length(x)
 
@@ -399,6 +399,59 @@ c.rlang_trace <- function(...) {
 
   new_trace(calls, parents, envs)
 }
+
+
+# Subsets sibling nodes, at the level of the rightmost leaf by
+# default. Supports full vector subsetting semantics (negative values,
+# missing index, etc).
+trace_subset_across <- function(trace, i, n = NULL) {
+  level <- trace_level(trace, n)
+  level_n <- length(level)
+  i <- validate_index(i, level_n)
+
+  indices <- unlist(map(level[i], chain_indices, trace$parents))
+  trace_subset(trace, indices)
+}
+trace_level <- function(trace, n = NULL) {
+  n <- n %||% trace_length(trace)
+  parents <- trace$parents
+  which(parents == parents[[n]])
+}
+
+chain_indices <- function(i, parents) {
+  c(
+    parents_indices(i, parents),
+    children_indices(i, parents)
+  )
+}
+children_indices <- function(i, parents) {
+  n <- length(parents)
+  age <- parents[[i]]
+  ages <- parents[1:n]
+
+  non_children <- parents <= age
+  non_children[seq(1, i)] <- FALSE
+  non_children <- which(non_children)
+
+  if (length(non_children)) {
+    end <- non_children[[1]] - 1
+  } else {
+    end <- n
+  }
+
+  seq2(i + 1L, end)
+}
+parents_indices <- function(i, parents) {
+  path <- int()
+
+  while (i != 0) {
+    path <- c(path, i)
+    i <- parents[i]
+  }
+
+  rev(path)
+}
+
 
 # Trimming ----------------------------------------------------------------